network-grpc: Reorg client async adaptors
In module client, move the future and stream adaptor types
to separate submodules to declutter the module code,
leaving the beef of the protocol implementation.

Namespacing of the public types allows shorter, more logically
organized names.
